QPPCT-442: move MockitoExtension to test commons
