+++ rest-api/src/test/java/gov/cms/qpp/conversion/api/config/AsyncConfigTest.java
@@ -2,25 +2,25 @@ package gov.cms.qpp.conversion.api.config;
 
 import gov.cms.qpp.conversion.api.model.Constants;
 import gov.cms.qpp.test.PropertiesTestSuite;
-import org.junit.Before;
-import org.junit.Test;
+import org.junit.jupiter.api.BeforeEach;
+import org.junit.jupiter.api.Test;
 import org.springframework.core.task.SyncTaskExecutor;
 import org.springframework.core.task.TaskExecutor;
 import org.springframework.scheduling.concurrent.ThreadPoolTaskExecutor;
 
 import static com.google.common.truth.Truth.assertWithMessage;
 
-public class AsyncConfigTest extends PropertiesTestSuite {
+class AsyncConfigTest extends PropertiesTestSuite {
 
 	private AsyncConfig config;
 
-	@Before
-	public void setup() {
+	@BeforeEach
+	void setup() {
 		config = new AsyncConfig();
 	}
 
 	@Test
-	public void testIsSyncIfPropertyIsPresent() {
+	void testIsSyncIfPropertyIsPresent() {
 		System.setProperty(Constants.USE_SYNC_EXECUTOR, "anything blah blah blah");
 		TaskExecutor executor = config.taskExecutor();
 		assertWithMessage("Executor should be SyncExecutor, was " + executor)
@@ -28,7 +28,7 @@ public class AsyncConfigTest extends PropertiesTestSuite {
 	}
 
 	@Test
-	public void testIsAsyncIfPropertyIsAbsent() {
+	void testIsAsyncIfPropertyIsAbsent() {
 		TaskExecutor executor = config.taskExecutor();
 		assertWithMessage("Executor should be ThreadPoolTaskExecutor, was " + executor)
 				.that(executor).isInstanceOf(ThreadPoolTaskExecutor.class);
