+++ test-commons/pom.xml
@@ -32,11 +32,16 @@
 			<version>1.1</version>
 		</dependency>
 
-		<!-- This must remain as junit 4 (non-vintage) until test-commons code uses junit 5 -->
 		<dependency>
-			<groupId>junit</groupId>
-			<artifactId>junit</artifactId>
-			<version>4.12</version>
+			<groupId>org.junit.jupiter</groupId>
+			<artifactId>junit-jupiter-engine</artifactId>
+			<version>5.0.1</version>
+		</dependency>
+
+		<dependency>
+			<groupId>org.mockito</groupId>
+			<artifactId>mockito-core</artifactId>
+			<version>2.9.0</version>
 		</dependency>
 	</dependencies>
