+++ test-commons/src/main/java/gov/cms/qpp/test/LoadTestSuite.java
@@ -1,13 +1,14 @@
 package gov.cms.qpp.test;
 
-import org.junit.BeforeClass;
+
+import org.junit.jupiter.api.BeforeAll;
 
 /**
  * Ensures that an environment variable or system property {@code runLoadTests} is present, or skips the tests in the class
  */
 public abstract class LoadTestSuite {
 
-	@BeforeClass
+	@BeforeAll
 	public static void ensureLoadTestsShouldRun() {
 		VariableDependantTestHelper.assumeIsPresent("runLoadTests");
 	}
