+++ test-commons/src/main/java/gov/cms/qpp/test/MockitoExtension.java
@@ -1,4 +1,4 @@
-package gov.cms.qpp.conversion.api.helper;
+package gov.cms.qpp.test;
 
 import static org.mockito.Mockito.mock;
