+++ test-commons/src/main/java/gov/cms/qpp/test/PropertiesTestSuite.java
@@ -1,21 +1,20 @@
 package gov.cms.qpp.test;
 
 import java.util.Properties;
-
-import org.junit.After;
-import org.junit.Before;
+import org.junit.jupiter.api.AfterEach;
+import org.junit.jupiter.api.BeforeEach;
 
 public class PropertiesTestSuite {
 
 	private Properties properties;
 
-	@Before
+	@BeforeEach
 	public final void saveProperties() {
 		properties = new Properties();
 		properties.putAll(System.getProperties());
 	}
 
-	@After
+	@AfterEach
 	public final void resetProperties() {
 		System.setProperties(properties);
 	}
